+++ components/scifio/src/ome/scifio/util/SignedByteBuffer.java
@@ -1,6 +1,5 @@
 package ome.scifio.util;
 
-
 import java.awt.image.DataBuffer;
 
 /**
@@ -46,7 +45,7 @@ public class SignedByteBuffer extends DataBuffer {
 
   /* @see java.awt.image.DataBuffer#getElem(int, int) */
   public int getElem(int bank, int i) {
-    return (int) bankData[bank][i + getOffsets()[bank]];
+    return bankData[bank][i + getOffsets()[bank]];
   }
 
   /* @see java.awt.image.DataBuffer#setElem(int, int) */
